requirements.txt
streamlit==1.22.0
python-docx==0.8.11
streamlit-authenticator==0.1.4
streamlit-option-menu==0.3.2
